use 32bit float instead of 64bit for laplacian for speed

+++ indi_allsky/flask/views.py
@@ -2295,7 +2295,7 @@ class JsonFocusView(JsonView):
         vl_start = time.time()
 
         ### determine variance of laplacian
-        blur_score = cv2.Laplacian(image_data, cv2.CV_64F).var()
+        blur_score = cv2.Laplacian(image_data, cv2.CV_32F).var()
         json_data['blur_score'] = float(blur_score)
 
         vl_elapsed_s = time.time() - vl_start
